refactor: fix natneg error

+++ src/Servers/NatNegotiation/src/Abstraction/BaseClass/ResponseBase.cs
@@ -8,8 +8,11 @@ namespace UniSpyServer.Servers.NatNegotiation.Abstraction.BaseClass
     {
         protected new RequestBase _request => (RequestBase)base._request;
         protected new ResultBase _result => (ResultBase)base._result;
-        public new byte[] SendingBuffer{ get => (byte[])base.SendingBuffer;
-            protected set => base.SendingBuffer = value; }
+        public new byte[] SendingBuffer
+        {
+            get => (byte[])base.SendingBuffer;
+            protected set => base.SendingBuffer = value;
+        }
         public ResponseBase(UniSpyLib.Abstraction.BaseClass.RequestBase request, UniSpyLib.Abstraction.BaseClass.ResultBase result) : base(request, result)
         {
         }
@@ -17,9 +20,9 @@ namespace UniSpyServer.Servers.NatNegotiation.Abstraction.BaseClass
         {
             List<byte> data = new List<byte>();
             data.AddRange(RequestBase.MagicData);
-            data.Add(_request.Version);
+            data.Add((byte)_request.Version);
             data.Add((byte)_result.PacketType);
-            data.AddRange(BitConverter.GetBytes(_request.Cookie));
+            data.AddRange(BitConverter.GetBytes((uint)_request.Cookie));
             SendingBuffer = data.ToArray();
         }
     }
